+++ psl-core/src/main/java/org/linqs/psl/model/rule/misc/GroundLinearConstraint.java
@@ -64,7 +64,7 @@ public class GroundLinearConstraint implements UnweightedGroundRule {
 
     @Override
     public ConstraintTerm getConstraintDefinition() {
-        GeneralFunction sum = new GeneralFunction(false, false, (short)atoms.length);
+        GeneralFunction sum = new GeneralFunction(false, false, atoms.length);
         for (int i = 0; i < atoms.length; i++) {
             sum.add(coeffs[i], atoms[i]);
         }
